requirements.txt
asgiref==3.4.1
Django==3.2.6
django-crispy-forms==1.12.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
